roleController单元测试

+++ src/main/java/com/mimi/cgims/web/controller/RoleController.java
@@ -50,6 +50,9 @@ public class RoleController {
         if (StringUtils.isNotBlank(error)) {
             return ResultUtil.getFailResultMap(error);
         }
+        if(Constants.ROLE_NAME_ADMIN.equals(role.getName())){
+            return ResultUtil.getFailResultMap("不能修改超级管理员角色");
+        }
         RoleModel newModel = roleService.get(id);
         BeanUtils.copyProperties(role, newModel, "id", "users", "permissions");
         newModel.setPermissions(buildPermissions(permissionIds));
@@ -74,7 +77,24 @@ public class RoleController {
     @RequestMapping(value = "/role/{id}", method = RequestMethod.DELETE)
     @ResponseBody
     public Object delete(@PathVariable String id) {
+        RoleModel role = roleService.getByName(Constants.ROLE_NAME_ADMIN);
+        if(id.contains(role.getId())){
+            return ResultUtil.getFailResultMap("不能删除超级管理员角色");
+        }
         roleService.delete(id);
         return ResultUtil.getSuccessResultMap(id);
+    }
+
+    @RequestMapping(value = "/role/batch", method = RequestMethod.POST)
+    @ResponseBody
+    public Object batch(String ids) {
+        if (StringUtils.isNotBlank(ids)) {
+            RoleModel role = roleService.getByName(Constants.ROLE_NAME_ADMIN);
+            if(ids.contains(role.getId())){
+                return ResultUtil.getFailResultMap("不能删除超级管理员角色");
+            }
+            roleService.batchDelete(ids.split(Constants.SPLIT_STRING_IDS));
+        }
+        return ResultUtil.getSuccessResultMap(ids);
     }
 }
